requirements.txt
streamlit>=1.30.0
google-genai>=1.0.0
pillow>=10.0.0
